requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.5
djangorestframework==3.16.1
drf-yasg==1.21.10
frozenlist==1.7.0
geoip2==5.1.0
idna==3.10
inflection==0.5.1
maxminddb==2.8.2
multidict==6.6.3
packaging==25.0
propcache==0.3.2
pytest~=8.4.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
